+++ client/pages/user/index.vue
@@ -7,6 +7,9 @@ if (error.value) {
 }
 
 async function deleteUser(id: string) {
+  if (!confirm('Are you sure you want to delete?')) {
+    return
+  }
   const { error } = await useMyFetch('/user', {
     method: 'delete',
     body: { id },
@@ -21,12 +24,12 @@ async function deleteUser(id: string) {
 
 <template>
   <span v-if="error">Couldn't fetch the users' data.</span>
-  <div v-else class='flex flex-col gap-y-[inherit]'>
+  <div v-else class="flex flex-col gap-y-[inherit]">
     <h1>Users</h1>
 
     <NuxtLink to="/user/create" class="button">Create a new user</NuxtLink>
 
-    <div v-for="{ id, name, email } in users" class="flex gap-x-4 items-center">
+    <div v-for="{ id, name, email } in users" class="flex gap-x-8 items-center">
       <span>{{ name }}</span>
       <span>{{ email }}</span>
       <NuxtLink :to="'/user/update/' + id">Edit</NuxtLink>
